# api/openapi.yaml
openapi: 3.0.3
info:
  title: Axiom owner-console release API
  version: 1.3.0
  description: >-
    Authenticated research, operations, audit, sandbox, and qualification
    control plane. Axiom V1 is Spot-only research software. Real trading,
    production-private order submission, and external asset movement are absent.
servers:
  - url: /
...
paths:
...
  /api/v1/session/login:
    post:
      operationId: loginSession
      tags: [Session]
      summary: Authenticate and rotate to a fresh opaque session
      parameters:
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "201":
          description: Authenticated session created
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/session/logout:
    post:
      operationId: logoutSession
      tags: [Session]
      summary: Revoke the current session and clear its cookies
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/session/me:
    get:
      operationId: getCurrentSession
      tags: [Session]
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current user and session projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionMe" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/run-catalog:
    get:
      operationId: getRunCatalog
      tags: [Research]
      summary: Return server-approved strategy and run choices
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Semantic run choices and plain blockers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunCatalog" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/demonstrations:
    get:
      operationId: listGuidedDemonstrations
      tags: [Research]
      summary: List executable synthetic guided demonstrations
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Server-owned executable synthetic walkthroughs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuidedDemonstrationPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/demonstrations/{id}:
    get:
      operationId: getGuidedDemonstration
      tags: [Research]
      summary: Execute and return one deterministic synthetic guided demonstration
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Canonical same-pipeline synthetic walkthrough result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuidedDemonstrationResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs:
    get:
      operationId: listRuns
      tags: [Research]
      summary: List durable owner-facing run records
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Durable runs in reverse creation order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
    post:
      operationId: createRun
      tags: [Research]
      summary: Start one server-resolved durable owner run
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RunCreateRequest" }
      responses:
        "202":
          description: A durable run was created from server-resolved immutable inputs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunResource" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/runs/{id}:
    get:
      operationId: getRun
      tags: [Research]
      summary: Get one owner-facing run record
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Semantic durable run projection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunResource" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/runs/{id}/timeline:
    get:
      operationId: getRunTimeline
      tags: [Research]
      summary: Read immutable event evidence for one run
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Ordered recorded run events
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunOutputPage" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/decisions:
    get:
      operationId: getRunDecisions
      tags: [Research]
      summary: Read immutable strategy decisions for one run
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Recorded decisions
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunOutputPage" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/orders:
    get:
      operationId: getRunOrders
      tags: [Research]
      summary: Read planned orders for one run
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Recorded planned orders
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunOutputPage" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/fills:
    get:
      operationId: getRunFills
      tags: [Research]
      summary: Read simulated execution events for one run
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Recorded simulated execution events
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunOutputPage" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/portfolio:
    get:
      operationId: getRunPortfolio
      tags: [Research]
      summary: Read the latest recorded virtual portfolio projection
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Latest recorded portfolio state, if the run has produced one
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/RunPortfolioProjection" },
                },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/risk:
    get:
      operationId: getRunRisk
      tags: [Research]
      summary: Read the run risk-evidence availability state
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Run-specific risk evidence availability
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunRiskProjection" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/runs/{id}/evidence:
    get:
      operationId: getRunEvidence
      tags: [Research]
      summary: Read safe immutable run identity and evidence hashes
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/ResourceID" }]
      responses:
        "200":
          description: Run evidence identity without storage paths or credentials
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RunEvidence" } },
            }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/data-catalogue:
    get:
      operationId: listDataCatalogue
      tags: [Research]
      summary: List protected server-registered data available to the owner
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Immutable data coverage and evidence identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataCataloguePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/stream:
    get:
      operationId: streamEvents
      tags: [Stream]
      summary: Resume durable operational events after a REST snapshot revision
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/Origin"
        - in: header
          name: Last-Event-ID
          schema: { $ref: "#/components/schemas/Revision" }
        - in: query
          name: after_revision
          schema: { $ref: "#/components/schemas/Revision" }
      responses:
        "200":
          description: Versioned SSE stream with 15-second heartbeats
          content:
            text/event-stream:
              schema:
                type: string
                description: Each data field is one StreamEvent JSON object.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "410":
          description: Resume cursor expired; fetch fresh REST snapshots
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: axiom_session
    operationalBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Origin:
      in: header
      name: Origin
      required: true
      schema: { type: string, format: uri, maxLength: 512 }
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 32, maxLength: 512 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        {
          type: string,
          minLength: 16,
          maxLength: 128,
          pattern: "^[A-Za-z0-9._:-]+$",
        }
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, minLength: 1, maxLength: 160 }
...
  responses:
    BadRequest:
      description: Request failed safe validation
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authorization, Origin, or CSRF policy denied the request
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Idempotency payload or lifecycle conflict
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    PreconditionFailed:
      description: Expected entity revision is stale
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Durable authentication or workload quota exceeded
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unavailable:
      description: A required fail-closed prerequisite is unavailable
      content:
        { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Decimal:
      type: string
      pattern: "^-?(0|[1-9][0-9]*)(\\.[0-9]+)?$"
      example: "1000.000000000000000000"
...
    Revision:
      type: string
      pattern: "^(0|[1-9][0-9]*)$"
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp with an explicit UTC offset.
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlation_id]
      properties:
        code: { type: string, minLength: 1, maxLength: 80 }
        message: { type: string, minLength: 1, maxLength: 240 }
        correlation_id: { type: string, minLength: 1, maxLength: 160 }
        summary: { type: string, minLength: 1, maxLength: 240 }
        detail: { type: string, minLength: 1, maxLength: 2000 }
        impact: { type: string, minLength: 1, maxLength: 500 }
        suggested_action: { type: string, minLength: 1, maxLength: 500 }
        blocking_prerequisites:
          type: array
          maxItems: 16
          items: { type: string, minLength: 1, maxLength: 240 }
        current_state: { type: string, minLength: 1, maxLength: 160 }
        required_state: { type: string, minLength: 1, maxLength: 160 }
        field_details:
          type: object
          additionalProperties: { type: string, maxLength: 160 }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, minLength: 3, maxLength: 254 }
        password:
          { type: string, minLength: 1, maxLength: 1024, writeOnly: true }
    LoginResponse:
      type: object
      additionalProperties: false
      required: [user, csrf_token, expires_at]
      properties:
        user: { $ref: "#/components/schemas/SessionUser" }
        csrf_token: { type: string, minLength: 32, maxLength: 512 }
        expires_at: { $ref: "#/components/schemas/Timestamp" }
    SessionUser:
      type: object
      additionalProperties: false
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
    SessionMe:
      type: object
      additionalProperties: false
      required: [user, session_id, session_revision, reauthenticated_at]
      properties:
        user: { $ref: "#/components/schemas/SessionUser" }
        session_id: { type: string }
        session_revision: { $ref: "#/components/schemas/Revision" }
        reauthenticated_at: { $ref: "#/components/schemas/Timestamp" }
    RunCatalog:
      type: object
      additionalProperties: false
      required: [choices]
      properties:
        choices:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/RunChoice" }
        blocker: { $ref: "#/components/schemas/RunBlocker" }
    GuidedDemonstrationPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/GuidedDemonstration" }
...
    GuidedDemonstrationResult:
      type: object
      additionalProperties: false
      required:
        [
          id,
          strategy_id,
          strategy_version,
          synthetic,
          advisory_only,
          configuration_hash,
          accepted,
          rejected,
          metrics,
          result_hash,
        ]
      properties:
        id: { type: string, minLength: 1 }
        strategy_id: { type: string, minLength: 1 }
        strategy_version: { type: string, minLength: 1 }
        synthetic: { type: boolean, const: true }
        advisory_only: { type: boolean }
        advisory_evidence:
          {
            type: string,
            minLength: 2,
            description: Canonical JSON advisory evidence; present only for non-order-capable demonstrations,
          }
        configuration_hash: { type: string, pattern: "^[a-f0-9]{64}$" }
        accepted: { $ref: "#/components/schemas/GuidedDemonstrationEvent" }
        rejected: { $ref: "#/components/schemas/GuidedDemonstrationEvent" }
        metrics:
          {
            type: string,
            minLength: 2,
            description: Canonical JSON metric payload; not profitability evidence,
          }
        result_hash: { type: string, pattern: "^[a-f0-9]{64}$" }
    RunChoice:
      type: object
      additionalProperties: false
      required:
        [
          strategy_id,
          strategy_name,
          strategy_version,
          mode,
          exchanges,
          instrument,
          cadence,
          warmup,
          order_capable,
        ]
      properties:
        strategy_id: { type: string }
        strategy_name: { type: string }
        strategy_version: { type: string }
        mode:
          {
            type: string,
            enum:
              [demonstration, backtest, replay, shadow, sandbox, testnet, demo],
          }
        exchanges:
          type: array
          minItems: 1
          maxItems: 2
          items: { type: string, enum: [binance, bybit] }
        instrument: { type: string }
        cadence: { type: string }
        warmup: { type: string }
        order_capable: { type: boolean }
    RunBlocker:
      type: object
      additionalProperties: false
      required: [code, summary, detail, suggested_action]
      properties:
        code: { type: string }
        summary: { type: string }
        detail: { type: string }
        suggested_action: { type: string }
    RunPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/RunResource" }
    RunResource:
      type: object
      additionalProperties: false
      required:
        [
          id,
          friendly_name,
          strategy_id,
          strategy_version,
          mode,
          environment,
          state,
          order_capable,
          available_actions,
          revision,
          created_at,
        ]
      properties:
        id: { type: string, minLength: 1 }
        friendly_name: { type: string, minLength: 1 }
        strategy_id: { type: string, minLength: 1 }
        strategy_version: { type: string, minLength: 1 }
        exchanges:
          type: array
          minItems: 1
          maxItems: 2
          uniqueItems: true
          items: { type: string, enum: [binance, bybit] }
        instrument: { type: string, minLength: 3, maxLength: 32 }
        mode: { type: string, enum: [backtest, replay, shadow, sandbox] }
        environment:
          {
            type: string,
            enum:
              [
                recorded_data,
                production_public,
                binance_spot_testnet,
                bybit_demo,
                paired_exchange_sandbox,
              ],
          }
        state: { type: string, minLength: 1 }
        order_capable: { type: boolean }
        available_actions:
          type: array
          maxItems: 4
          uniqueItems: true
          items: { $ref: "#/components/schemas/RunAction" }
        revision: { $ref: "#/components/schemas/Revision" }
        waiting_reason: { type: string, maxLength: 500 }
        next_evaluation_at: { $ref: "#/components/schemas/Timestamp" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
    RunAction:
      type: string
      enum: [pause, resume, step, stop]
...
    RunOutputPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 500
          items: { $ref: "#/components/schemas/RunOutput" }
    RunPortfolioProjection:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state: { type: string, enum: [recorded, not_recorded] }
        summary: { type: string, minLength: 1, maxLength: 500 }
        realized_pnl: { $ref: "#/components/schemas/Decimal" }
        unrealized_pnl: { $ref: "#/components/schemas/Decimal" }
        total_pnl: { $ref: "#/components/schemas/Decimal" }
        account_drawdown: { $ref: "#/components/schemas/NonnegativeDecimal" }
        slippage: { $ref: "#/components/schemas/NonnegativeDecimal" }
        positions:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/RunPortfolioPosition" }
        fees:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/RunPortfolioFee" }
        ordinal: { $ref: "#/components/schemas/Revision" }
        content_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        canonical_payload: { type: string, minLength: 2, maxLength: 1048576 }
        waiting_reason: { type: string, maxLength: 500 }
...
    RunRiskProjection:
      type: object
      additionalProperties: false
      required: [state, summary]
      properties:
        state:
          type: string
          enum: [recorded, not_recorded]
          x-enum-varnames:
            [RunRiskProjectionStateRecorded, RunRiskProjectionStateNotRecorded]
        summary: { type: string, maxLength: 500 }
        status:
          type: string
          enum: [normal, waiting, blocked]
          x-enum-varnames:
            [
              RunRiskProjectionStatusNormal,
              RunRiskProjectionStatusWaiting,
              RunRiskProjectionStatusBlocked,
            ]
        blockers:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 160 }
        observations:
          type: array
          maxItems: 20
          items: { $ref: "#/components/schemas/RunRiskObservation" }
...
    RunEvidence:
      type: object
      additionalProperties: false
      required: [state]
      properties:
        state: { type: string, enum: [recorded, not_recorded] }
        manifest_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        source_commit: { type: string, pattern: "^[0-9a-f]{40,64}$" }
        configuration_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        dataset_manifest_hash: { type: string, pattern: "^[0-9a-f]{64}$" }
        model_namespace: { type: string, minLength: 1, maxLength: 200 }
        confidence_tier: { type: string, enum: [A, B, C, D] }
    RunCreateRequest:
      type: object
      additionalProperties: false
      required:
        [strategy_id, strategy_version, mode, exchanges, instrument, preset]
      properties:
        strategy_id:
          { type: string, minLength: 3, maxLength: 64, pattern: "^[a-z0-9-]+$" }
        strategy_version: { type: string, minLength: 3, maxLength: 128 }
        mode:
          {
            type: string,
            enum:
              [backtest, replay, shadow, sandbox, testnet, demo, demonstration],
          }
        exchanges:
          type: array
          minItems: 1
          maxItems: 2
          uniqueItems: true
          items: { type: string, enum: [binance, bybit] }
        instrument: { type: string, minLength: 3, maxLength: 32 }
        preset: { type: string, enum: [latest-qualified-inputs] }
    DataCataloguePage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/DataCatalogueItem" }
...